share more block code
